% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{allscenarios}
\alias{allscenarios}
\title{Make a tibble of all scenarios}
\usage{
allscenarios()
}
\description{
a Scenario is defined by three entries:
\enumerate{
\item the climate model
\item the RCP definition
\item the time period
}
}
\details{
This function returns all possible scenarios.
}
\examples{
allscenarios()
}
