% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{gaez_download_yield_allcrops}
\alias{gaez_download_yield_allcrops}
\title{Batch download All Crops}
\usage{
gaez_download_yield_allcrops(dir = ".", scenarios = allscenarios())
}
\description{
Downloads all crops for \code{\link[=allscenarios]{allscenarios()}}
}
