% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gaez_filename}
\alias{gaez_filename}
\title{Gaez Filename Constructor}
\usage{
gaez_filename(crop, variable = "yl", input = "H", irrigation = "r", co2 = "0")
}
\description{
Gaez Filename Constructor
}
\examples{
gaez_filename("popeye-spinach","yx","L","r","1")
}
