% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{gaez_LR_download}
\alias{gaez_LR_download}
\title{Land and Water Resources Download}
\usage{
gaez_LR_download(dir = ".", subtheme = "soi1", variable = "hwsd_domi_30s")
}
\description{
Same as \code{\link[=gaez_download]{gaez_download()}} but for the Land and Water Resources Theme.
}
\details{
Notice that this vectorizes naturally for a vector of variable names within the same subtheme.
}
\examples{
terrain_vars = terrain_resources_varnames()[["variable"]]
length(terrain_vars)
gaez_LR_download(subtheme = "ter",variable = terrain_vars)
}
