% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAEZr-package.R
\docType{package}
\name{GAEZr-package}
\alias{GAEZr}
\alias{GAEZr-package}
\title{GAEZr: Download and Process GAEZ v4 Data}
\description{
Helps to download GAEZ v4 data directly into R from web server.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://floswald.github.io/GAEZr/}
}

}
\author{
\strong{Maintainer}: Florian Oswald \email{florian.oswald@gmail.com} (\href{https://orcid.org/0000-0001-7737-2038}{ORCID})

Other contributors:
\itemize{
  \item Christophe Gouel \email{christophe.gouel@inrae.fr} (\href{https://orcid.org/0000-0002-3946-5222}{ORCID}) [contributor]
}

}
\keyword{internal}
