% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{soil_resources_varnames}
\alias{soil_resources_varnames}
\title{Variable Names for Soil Resources}
\usage{
soil_resources_varnames()
}
\description{
This is a subtheme of the \emph{Land and Water Resources} theme of GAEZ. Part of the \code{soi1} subtheme.
}
\examples{
soil_resources_varnames()
}
