% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{soil_suitability_varnames}
\alias{soil_suitability_varnames}
\title{Variable Names for Soil Suitability}
\usage{
soil_suitability_varnames()
}
\description{
This is a subtheme of the \emph{Land and Water Resources} theme of GAEZ. Part of the \code{soi2} subtheme.
}
\examples{
soil_suitability_varnames()
}
