% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{terrain_resources_varnames}
\alias{terrain_resources_varnames}
\title{Variable Names for Terrain Resources}
\usage{
terrain_resources_varnames()
}
\description{
This is a subtheme of the \emph{Land and Water Resources} theme of GAEZ. Part of the \code{ter} subtheme.
}
\examples{
terrain_resources_varnames()
}
