% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme1_subs}
\alias{theme1_subs}
\title{Subtheme Names for Theme 1 of GAEZ v4}
\usage{
theme1_subs()
}
\description{
This are the subthemes of \emph{Land and Water Resources} theme of GAEZ.
}
\examples{
theme1_subs()
}
