% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme4_varnames}
\alias{theme4_varnames}
\title{Variable Names for Theme 4 of GAEZ v4}
\usage{
theme4_varnames()
}
\description{
This is the \emph{Suitability and Attainable Yield} theme of GAEZ.
}
\examples{
theme4_varnames()
}
